+++ vagrant/mysql.sh
@@ -1,8 +1,15 @@
-apt-get install debconf-utils -y
+#!/usr/bin/env bash
 
-debconf-set-selections <<< "mysql-server mysql-server/root_password password P@ssw0rd2"
-debconf-set-selections <<< "mysql-server mysql-server/root_password_again password P@ssw0rd2"
+if [ ! -f /tmp/provision.mysql ]; then
 
-apt-get install mysql-server -y
+  apt-get install debconf-utils -y
 
-echo "CREATE DATABASE IF NOT EXISTS movies" | mysql -u root --password=P@ssw0rd2
+  debconf-set-selections <<< "mysql-server mysql-server/root_password password P@ssw0rd2"
+  debconf-set-selections <<< "mysql-server mysql-server/root_password_again password P@ssw0rd2"
+
+  apt-get install mysql-server -y
+
+  echo "CREATE DATABASE IF NOT EXISTS movies" | mysql -u root --password=P@ssw0rd2
+
+  touch /tmp/provision.mysql
+fi
